операции с деньгами - BigDecimal

+++ marketplace-pricing-service/src/main/java/org/yrti/pricing/dto/PricingResponse.java
@@ -3,10 +3,12 @@ package org.yrti.pricing.dto;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 @AllArgsConstructor
 public class PricingResponse {
     private Long productId;
-    private Double originalPrice;
-    private Double discountedPrice;
+    private BigDecimal originalPrice;
+    private BigDecimal discountedPrice;
 }
